permute_alleles(): a new function which permutes alleles at all but the last locus. -----------------
main_proc():
-----------
A loop has been added that begins just after the declarations and ends just before
the end of main_proc(). If permu_flag is set to 1, then MAX_PERMU permutations are
done. Otherwise, only the 0th permutation (the origninal data with no permuting)
is executed. This flag is currently set within main_proc(), but should be read in
as an iput parameter.

The following are initialized before each permutation
  pheno[0]
  geno[i][j]
  genopheno[i][j]
  allele_freq[j][i]

Two files are opened for output
  summary_iter.out  contains a summary of the iterations for the 0th permutation
  summary_permu.out contains a summary of the LLs and LRs for the permutations