main_proc():
The id_unique_alleles() function is now only called when permu=0.

The initialization of the genopheno[][] array was removed from the
"pre-processing for permutations". This was one of the more time
intensive spots from Alex's profiling. The value inserted in genopheno[][]
now is related to the permutation number (instead of a 0/1). So,
genopheno[i][j]==1 is now replaced by genopheno[i][j]>permu

Computations for Ken Kidd's normalized LD statistic were added. This is
the LR statistic normalized by the mean and standard deviation of the
permutation distribution for the LR statistic. The statistic, mean and SD
of the permutation distribution are only computed and printed when the
'-p' option is selected.

emcalc():

hap_freq[i][iter] was changed to hap_freq[i] since only the previous
iteration is used in the iterations. hap_freq[] is now in the argument
list for the function call (rather than freq_zero[], which was then
copied to hap_freq[i][0] previously). This also avoids the need for
  static double hap_freq[MAX_HAPLOS][MAX_ITER];
  INIT_STATIC_DIM2(double, hap_freq, MAX_HAPLOS, MAX_ITER);

The loglike computation was moved into the loop for computing pheno freqs
to avoid an extra loop through all phenos.