* Main.py (Main._runFilters): Parse the ini file for "directory" and "file" rather than "path." Change the way CustomBinning is called to account for changes in that method.
* Filter.py (BinningFilter.__init__): Receive the binningPath when
this object is instatiated.
(BinningFilter.doCustomBinning): Modularized this method so that
it can be called in line with the other filters.  Method now looks
for custom binning specifications in a single file, specified in
the CustomBinning section of the ini file using the file option.

* config.ini: For greater user clarity, paths which are
directories are now indicated by the option "directory" and paths
which are files by "file."