Add a Singularity bootstrap, based on Fedora 24
As some of pypop's dependencies have become obsolete, with the exception of
on Fedora 24 (where things like the Numeric Python module are still
available), one way of distributing pypop is as a container!

This adds a Singularity container bootstrap definition file.  The bootstrap
is configured to install a minimal 64-bit CentOS 24 image, along with the
various required Python modules, plus third-party libraries, plus a
development environment.  pypop's source is then copied into the container,
and built.

A basic "run script" is installed, so that when a user attempts to run the
container, it runs pypop directly.  All command-line options are passed to
pypop.

The `yum`, `rpm`, `vim`, and `less` packages are also installed, even
though they aren't required, so that (if needed) a user can shell into the
container and do basic work, assuming...

* The container is shelled into in "write mode".
* The user is running this as root (since everything in the container is
owned by root).