Singularity: Disable diffs
As we already know, the diffs are showing differences between expected
and actual output.  That's OK right now, and I added "|| true" to the
end of each diff to ensure that the shell exit code would still be
zero.

Yet, it seems like that's not happening.  Looking at a recent build
on shub, I am seeing the first diff happen, but the second diff
never starts, and the build fails.

I'm thinking that something's happening to ignore the "|| true" suffix.
So, for now, let's take them out.