Singularity: Move to ubuntu base, and build SWIG
This moves us to the Ubuntu 17.04 base.

Unfortunately, the SWIG included with ubuntu has issues.
So, we download and build in SWIG 3.0.12 ourselves.

Finally, we do some post-building cleanup, removing the
build tools and -dev packages, as well as SWIG source.
It won't save any space now, but (once Singularity supports
image shrinking) it eventually will!